requirements.txt
streamlit==1.37.1
openai==1.40.0
python-dotenv==1.0.1
rank-bm25==0.2.2
numpy==1.26.4
pandas==2.2.2
altair==5.3.0
plotly==5.23.0
faiss-cpu==1.7.4
tqdm==4.66.4
scikit-learn==1.4.2
ipykernel==6.25.2
rouge-score==0.1.2
httpx==0.27.2